Refactor : js, html 파일 분리 profile group 수정

+++ Core/src/main/java/com/example/core/dto/OrderDto.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 public class OrderDto {
     private String storeName;
     private List<MenuOrderDTO> menuOrders;
-    private int LoginId;
+    private Long LoginId;
 
     @Getter
     @Setter
