+++ DeliveryAPI/src/main/java/com/example/deliveryapi/model/OrderDetailRepository.java
@@ -3,5 +3,5 @@ package com.example.deliveryapi.model;
 import com.example.deliveryapi.entity.OrderDetailDataEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderDetailRepository extends JpaRepository<OrderDetailDataEntity, Integer> {
+public interface OrderDetailRepository extends JpaRepository<OrderDetailDataEntity, Long> {
 }
