+++ DeliveryAPI/src/main/java/com/example/deliveryapi/model/OrderRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderRepository extends JpaRepository<OrderDataEntity, Integer> {
+public interface OrderRepository extends JpaRepository<OrderDataEntity, Long> {
     List<OrderDataEntity> findByUserUserIdAndOrderDatetimeBetween(long userId, LocalDate start, LocalDate end);
     List<OrderDataEntity> findByStoreUserIdAndOrderDatetimeBetween(long storeId, LocalDate start, LocalDate end);
 }
