+++ DeliveryAPI/src/main/java/com/example/deliveryapi/model/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserSignupDataEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserSignupDataEntity, Long> {
     UserSignupDataEntity findByUserName(String username);
     UserSignupDataEntity findByCompanyName(String companyName);
